Fix sign-conversion warning in BitReader::ReadBits
- Cast bit offset calculation to std::size_t before arithmetic
- Resolves -Wsign-conversion error when updating position

+++ applications/src/TestBattle/BitReader.hpp
@@ -53,7 +53,7 @@ public:
     temp                     = (temp >> _bitPosition) & mask;
 
     // Update the position and bit position within the data vector
-    position                 = current_position + (count + _bitPosition) / bit8;
+    position                 = current_position + static_cast<size_t>(count + _bitPosition) / bit8;
     _bitPosition             = (count + _bitPosition) % bit8;
 
     // Return the extracted bits as a std::uint16_t
